Altered interactions for launching bobber to work properly with animations and effects.
