+++ core/src/main/java/dev/rm20/anglersalmanac/MinigameManager/MinigameManager.java
@@ -29,7 +29,7 @@ import dev.rm20.anglersalmanac.AnglersAlmanac;
 import dev.rm20.anglersalmanac.Components.BobberComponent;
 import dev.rm20.anglersalmanac.Components.MinigameComponent_TensionBar;
 import dev.rm20.anglersalmanac.IEvents.LootCaughtEvent;
-import dev.rm20.anglersalmanac.Interactions.LaunchBobberInteraction;
+import dev.rm20.anglersalmanac.Interactions.Rod.UseRodInteraction;
 import dev.rm20.anglersalmanac.Metadata.FishingContext;
 import dev.rm20.anglersalmanac.Metadata.FishingModifier;
 import dev.rm20.anglersalmanac.Metadata.FishingRodData;
@@ -66,20 +66,20 @@ public class MinigameManager {
             case "TensionBar":
                 FishingRodData meta = fishingRod.getFromMetadataOrNull(FishingRodData.KEYED_CODEC);
                 if (meta == null) {
-                    LaunchBobberInteraction.cancelFishing(commandBuffer, player, fishingRod);
+                    UseRodInteraction.cancelFishing(commandBuffer, player, fishingRod);
                     break;
                 }
                 MinigameComponent_TensionBar minigame = MinigameComponent_TensionBar.spawnMinigame(commandBuffer, player.getReference(), bobberRef, fishingRod.getItemId());
-                LaunchBobberInteraction.updateMetadata(hotbarComp, hotbarComp.getActiveSlot(), hotbarComp.getActiveItem(), meta.getBoundBobber(), minigame.selfUUID, 1);
+                UseRodInteraction.updateMetadata(hotbarComp, hotbarComp.getActiveSlot(), hotbarComp.getActiveItem(), meta.getBoundBobber(), minigame.selfUUID, 1);
 
                 break;
             case "NoMinigame":
                 DropLoot(FirstRoll(bobberRef, player, commandBuffer, depth), player, commandBuffer, bobberRef, -1);
-                LaunchBobberInteraction.cancelFishing(commandBuffer, player, fishingRod);
+                UseRodInteraction.cancelFishing(commandBuffer, player, fishingRod);
                 break;
             default: // No Minigame, just reel fish.
                 DropLoot(FirstRoll(bobberRef, player, commandBuffer, depth), player, commandBuffer, bobberRef, -1);
-                LaunchBobberInteraction.cancelFishing(commandBuffer, player, fishingRod);
+                UseRodInteraction.cancelFishing(commandBuffer, player, fishingRod);
                 break;
         }
