+++ core/src/main/resources/Server/Item/Animations/AA_Rod.json
@@ -6,6 +6,11 @@
       "FirstPerson": "Characters/Animations/AA_Rod_Cast_FP.blockyanim",
       "Speed": 2,
       "BlendingDuration": 0.05
+    },
+    "Reel": {
+      "ThirdPerson": "Characters/Animations/Items/Main_Handed/Handgun/Attacks/Shoot/Shoot.blockyanim",
+      "FirstPerson": "Characters/Animations/Items/Main_Handed/Handgun/Attacks/Shoot/Shoot_FPS.blockyanim",
+      "Speed": 0.5
     }
   }
 }
