--- core/src/main/resources/Server/Item/Items/AA_DeleteMe.json
@@ -1,115 +1,0 @@
-{
-  "TranslationProperties": {
-    "Name": "server.items.Tool_Pickaxe_Cobalt.name"
-  },
-  "Parent": "Tool_Pickaxe_Crude",
-  "Icon": "Icons/ItemsGenerated/Tool_Pickaxe_Cobalt.png",
-  "Quality": "Rare",
-  "ItemLevel": 35,
-  "Model": "Items/Tools/Pickaxe/Cobalt.blockymodel",
-  "Texture": "Items/Tools/Pickaxe/Cobalt_Texture.png",
-  "Recipe": {
-    "TimeSeconds": 4,
-    "Input": [
-      {
-        "ItemId": "Ingredient_Bar_Cobalt",
-        "Quantity": 7
-      },
-      {
-        "ItemId": "Ingredient_Leather_Heavy",
-        "Quantity": 2
-      },
-      {
-        "ItemId": "Ingredient_Fabric_Scrap_Shadoweave",
-        "Quantity": 4
-      }
-    ],
-    "BenchRequirement": [
-      {
-        "Id": "Workbench",
-        "Type": "Crafting",
-        "Categories": [
-          "Workbench_Tools"
-        ],
-        "RequiredTierLevel": 2
-      }
-    ]
-  },
-  "Tool": {
-    "Specs": [
-      {
-        "Power": 1,
-        "GatherType": "SoftBlocks"
-      },
-      {
-        "Power": 0.5,
-        "GatherType": "Soils"
-      },
-      {
-        "Power": 0.05,
-        "GatherType": "Woods"
-      },
-      {
-        "Power": 0.5,
-        "GatherType": "Rocks",
-        "Quality": 4
-      },
-      {
-        "Power": 0.5,
-        "GatherType": "Benches"
-      },
-      {
-        "Power": 0.17,
-        "GatherType": "VolcanicRocks"
-      },
-      {
-        "Power": 0.5,
-        "GatherType": "OreCopper"
-      },
-      {
-        "Power": 0.5,
-        "GatherType": "OreIron"
-      },
-      {
-        "Power": 0.5,
-        "GatherType": "OreSilver"
-      },
-      {
-        "Power": 0.5,
-        "GatherType": "OreGold"
-      },
-      {
-        "Power": 0.25,
-        "GatherType": "OreThorium"
-      },
-      {
-        "Power": 0.25,
-        "GatherType": "OreCobalt"
-      },
-      {
-        "Power": 0.125,
-        "GatherType": "OreAdamantite",
-        "Quality": 4
-      },
-      {
-        "Power": 0.084,
-        "GatherType": "OreMithril"
-      }
-    ],
-    "DurabilityLossBlockTypes": [
-      {
-        "BlockSets": [
-          "Stone",
-          "Rock",
-          "Ores",
-          "Soil",
-          "Wood"
-        ],
-        "DurabilityLossOnHit": 0.25
-      }
-    ]
-  },
-  "MaxDurability": 325,
-  "PlayerAnimationsId": "AA_Rod",
-  "UsePlayerAnimations": false
-}
