--- core/src/main/resources/Server/Item/RootInteractions/AA_Rod_Cast.json
@@ -1,10 +1,0 @@
-{
-  "Interactions": [
-    "LaunchBobberInteraction"
-  ],
-  "Tags": {
-    "Attack": [
-      "Melee"
-    ]
-  }
-}
